backend/requirements.txt
flask
flask-cors
google-generativeai
pandas
python-dotenv